Sending (#32)
* Remove state comparison before MQTT publish

Eliminated the use of lastState and lastFveData to compare changes before publishing to MQTT. Now, heater and FVE data are always published when their respective functions are called, simplifying the logic and ensuring data is sent on every invocation. Also updated MQTT reconnection logic to use exponential backoff.

* Fix FVE data publishing with correct buffer initialization

Added a memcpy to copy currentFveData into the buffer before publishing to ensure the correct data is sent to Home Assistant.